requirements.txt
yt-dlp>=2023.11.16
moviepy>=1.0.3
inquirer>=3.1.3
google-api-python-client>=2.118.0
python-dotenv>=1.0.0
